Remove --no-clobber option
`wget --no-clobber` resulted in trouble for me. If the file exists, it will write "File ‘obj_19_pose_estimator_model-epoch199.pt’ already there; not retrieving." to the file specified after `-o`. The pipeline will then yield all kinds of cryptic errors, because it is trying to interpret this text as a model.